Move ForeignToplevelManagerV1Global definition from header to cpp
