+++ src/server/frontend_wayland/foreign_toplevel_manager_v1.h
@@ -19,7 +19,8 @@
 #ifndef MIR_FRONTEND_FOREIGN_TOPLEVEL_MANAGER_V1_H
 #define MIR_FRONTEND_FOREIGN_TOPLEVEL_MANAGER_V1_H
 
-#include "wlr-foreign-toplevel-management-unstable-v1_wrapper.h"
+#include <memory>
+struct wl_display;
 
 namespace mir
 {
@@ -31,27 +32,13 @@ class Shell;
 namespace frontend
 {
 class SurfaceStack;
+class ForeignToplevelManagerV1Global;
 
-/// Informs a client about toplevels from itself and other clients
-/// The Wayland objects it creates for each toplevel can be used to aquire information and control that toplevel
-/// Useful for task bars and app switchers
-class ForeignToplevelManagerV1Global
-    : public wayland::ForeignToplevelManagerV1::Global
-{
-public:
-    ForeignToplevelManagerV1Global(
-        wl_display* display,
-        std::shared_ptr<shell::Shell> shell,
-        std::shared_ptr<Executor> const& wayland_executor,
-        std::shared_ptr<SurfaceStack> const& surface_stack);
-
-    std::shared_ptr<shell::Shell> const shell;
-    std::shared_ptr<Executor> const wayland_executor;
-    std::shared_ptr<SurfaceStack> const surface_stack;
-
-private:
-    void bind(wl_resource* new_resource) override;
-};
+auto create_foreign_toplevel_manager_v1(
+    wl_display* display,
+    std::shared_ptr<shell::Shell> const& shell,
+    std::shared_ptr<Executor> const& wayland_executor,
+    std::shared_ptr<SurfaceStack> const& surface_stack) -> std::shared_ptr<ForeignToplevelManagerV1Global>;
 }
 }
